backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
altair==6.0.0
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.2
asttokens==3.0.0
attrs==25.4.0
bcrypt==4.0.1
beautifulsoup4==4.13.5
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
cmdstanpy==1.3.0
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.15
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
executing==2.2.0
fastapi==0.117.1
fonttools==4.61.1
fpdf2==2.8.5
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.46
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.183.0
google-auth==2.47.0
google-auth-httplib2==0.2.0
google-genai==1.59.0
googleapis-common-protos==1.70.0
grpcio==1.75.0
grpcio-status==1.71.2
h11==0.16.0
holidays==0.88
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.1.7
multidict==6.7.0
narwhals==2.15.0
nest-asyncio==1.6.0
numpy==2.3.3
openai==2.15.0
packaging==25.0
pandas==2.3.3
parso==0.8.4
passlib==1.7.4
pillow==12.1.0
platformdirs==4.3.8
plotly==6.5.2
prompt_toolkit==3.0.51
propcache==0.4.1
prophet==1.2.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.16.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.21
pytz==2025.2
pywin32==311
pyzmq==27.0.1
RapidFuzz==3.14.3
referencing==0.37.0
reportlab==4.4.9
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.17.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
stanio==0.5.1
starlette==0.48.0
streamlit==1.52.2
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
twilio==9.9.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.37.0
watchdog==6.0.0
wcwidth==0.2.13
websockets==15.0.1
xgboost==3.1.3
yarl==1.22.0
